pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vmware2scw"
version = "1.0.0"
description = "VMware to Scaleway Instance migration tool"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "vmware2scw contributors"},
]
keywords = ["vmware", "scaleway", "migration", "kvm", "qcow2"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "pyvmomi>=8.0.3",
    "boto3>=1.34",
    "rich>=13.7",
    "click>=8.1",
    "pyyaml>=6.0",
    "requests>=2.31",
    "jinja2>=3.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
vmware2scw = "vmware2scw.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
